Only on mac

+++ git/install.sh
@@ -1,2 +1,5 @@
-echo "pinentry-program /usr/local/bin/pinentry-mac" >> ~/.gnupg/gpg-agent.conf
-killall gpg-agent
+if [ "$(uname -s)" == "Darwin" ]
+then
+  echo "pinentry-program /usr/local/bin/pinentry-mac" >> ~/.gnupg/gpg-agent.conf
+  killall gpg-agent
+fi
